Moving to class based version
